porting draft

+++ source/SharedCrtResourceManager.cpp
@@ -317,7 +317,7 @@ int SharedCrtResourceManager::establishConnection(const PlainConfig &config)
     if (proxyConfig.httpProxyEnabled)
     {
         proxyOptions.HostName = proxyConfig.proxyHost->c_str();
-        proxyOptions.Port = proxyConfig.proxyPort.value();
+        proxyOptions.Port = static_cast<uint16_t>(proxyConfig.proxyPort.value());
 
         LOGM_INFO(
             TAG,
